forge fmt

+++ src/functions-type/Func.sol
@@ -3,16 +3,14 @@ pragma solidity ^0.8.29;
 
 contract C layout at 42 {
     uint256 private number;
+
     event Success();
 
     /// Declaration of a variable whose type is a function type
     function(address) external returns (bool) whenZero;
 
     /// A function with a function return type
-    function changeFunc()
-        public
-        returns (function(address) external returns (bool))
-    {
+    function changeFunc() public returns (function(address) external returns (bool)) {
         if (block.timestamp % 5 == 0) {
             whenZero = this.isZero;
             return whenZero;
@@ -23,10 +21,7 @@ contract C layout at 42 {
     }
 
     /// Change function selector assembly
-    function changeFuncAssembly(
-        address newAddr,
-        uint newSelector
-    ) public view returns (function() external func) {
+    function changeFuncAssembly(address newAddr, uint256 newSelector) public view returns (function() external func) {
         assembly {
             func.selector := newSelector
             func.address := newAddr
